refactor:  Secure AI palette generation with API route
